spi: zynq_qspi: Remove compilation warning caused by min macro
This should be the part of:
"zynq_qspi.c: Set correct frequency regardless of hardware setting"
(sha1: 879358544f7aee5d94c37d2a78622a4b783b8198)

Compilation warning:
In file included from include/common.h:26:0,
                 from drivers/spi/zynq_qspi.c:9:
drivers/spi/zynq_qspi.c: In function 'spi_setup_slave':
include/linux/kernel.h:155:17: warning: comparison of distinct pointer
types lacks a cast [enabled by default]
  (void) (&_min1 == &_min2);  \
                 ^
drivers/spi/zynq_qspi.c:822:28: note: in expansion of macro 'min'
  qspi->qspi.max_speed_hz = min(max_hz, lqspi_frequency);
                            ^
In file included from include/common.h:26:0,
                 from drivers/spi/zynq_qspi.c:9:
drivers/spi/zynq_qspi.c: In function 'spi_setup_slave':
include/linux/kernel.h:155:17: warning: comparison of distinct pointer
types lacks a cast [enabled by default]
  (void) (&_min1 == &_min2);  \
                 ^
drivers/spi/zynq_qspi.c:822:28: note: in expansion of macro 'min'
  qspi->qspi.max_speed_hz = min(max_hz, lqspi_frequency);

+++ drivers/spi/zynq_qspi.c
@@ -780,7 +780,7 @@ struct spi_slave *spi_setup_slave(unsigned int bus, unsigned int cs,
 {
 	int is_dual;
 	int is_dio = 0;
-	unsigned long lqspi_frequency;
+	unsigned int lqspi_frequency;
 	struct zynq_qspi_slave *qspi;
 
 	debug("%s: bus: %d cs: %d max_hz: %d mode: %d\n",
@@ -811,7 +811,7 @@ struct spi_slave *spi_setup_slave(unsigned int bus, unsigned int cs,
 		qspi->qspi.master.input_clk_hz = 200000000;
 	} else {
 		qspi->qspi.master.input_clk_hz = lqspi_frequency;
-		debug("Qspi clk frequency set to %ld Hz\n", lqspi_frequency);
+		debug("Qspi clk frequency set to %d Hz\n", lqspi_frequency);
 	}
 
 	qspi->slave.option = is_dual;
